# shared/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Qattan Curtains showroom API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
}

export type AdminUserOrNull = AdminUser | null;

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  nameEn: string | null;
  slug: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  imageUrl: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
  /** @nullable */
  nameEn?: string | null;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  nameEn: string | null;
  slug: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  specs: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  priceText: string | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  images: string[];
  isFeatured: boolean;
  isAvailable: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  /** @nullable */
  nameEn?: string | null;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  priceText?: string | null;
  /** @nullable */
  categoryId?: number | null;
  images?: string[];
  isFeatured?: boolean;
  isAvailable?: boolean;
}

export interface ProductListResponse {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface GalleryItem {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  imageUrl: string;
  sortOrder: number;
  createdAt: string;
}

export interface GalleryInput {
  title: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  sortOrder?: number;
}

export type InquiryStatus = (typeof InquiryStatus)[keyof typeof InquiryStatus];

export const InquiryStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface InquiryItem {
  productId: number;
  productName: string;
  quantity: number;
}

export interface Inquiry {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  message: string | null;
  items: InquiryItem[];
  status: InquiryStatus;
  createdAt: string;
}

export interface InquiryInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  message?: string | null;
  items?: InquiryItem[];
}

export type InquiryUpdateStatus =
  (typeof InquiryUpdateStatus)[keyof typeof InquiryUpdateStatus];

export const InquiryUpdateStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface InquiryUpdate {
  status?: InquiryUpdateStatus;
}

export interface Settings {
  phone: string;
  whatsapp: string;
  email: string;
  address: string;
  mapEmbedUrl: string;
  instagram: string;
  snapchat: string;
  twitter: string;
  brandTagline: string;
  heroTitle: string;
  heroSubtitle: string;
  aboutText: string;
}

export interface SettingsInput {
  phone?: string;
  whatsapp?: string;
  email?: string;
  address?: string;
  mapEmbedUrl?: string;
  instagram?: string;
  snapchat?: string;
  twitter?: string;
  brandTagline?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  aboutText?: string;
}

export type AdminStatsProductsByCategoryItem = {
  categoryId: number;
  categoryName: string;
  count: number;
};

export interface AdminStats {
  productsCount: number;
  categoriesCount: number;
  galleryCount: number;
  inquiriesCount: number;
  newInquiriesCount: number;
  featuredCount: number;
  recentInquiries: Inquiry[];
  productsByCategory: AdminStatsProductsByCategoryItem[];
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListProductsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
  page?: number;
  limit?: number;
};