# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Al Qattan Curtains showroom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      operationId: createCategory
      tags: [categories, admin]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
    delete:
      operationId: deleteCategory
      tags: [categories, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /categories/by-slug/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GalleryItem" }
    post:
      operationId: createGalleryItem
      tags: [gallery, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GalleryInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GalleryItem" }
  /gallery/{id}:
    patch:
      operationId: updateGalleryItem
      tags: [gallery, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GalleryInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GalleryItem" }
    delete:
      operationId: deleteGalleryItem
      tags: [gallery, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /inquiries:
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit an inquiry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InquiryInput" }
      responses:
        "201":
          description: Inquiry created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Inquiry" }
    get:
      operationId: listInquiries
      tags: [inquiries, admin]
      summary: List inquiries (admin)
      responses:
        "200":
          description: Inquiries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Inquiry" }
  /inquiries/{id}:
    patch:
      operationId: updateInquiry
      tags: [inquiries, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InquiryUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Inquiry" }
    delete:
      operationId: deleteInquiry
      tags: [inquiries, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AdminUser:
      type: object
      required: [id, email, name]
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
...
    Category:
      type: object
      required: [id, name, nameEn, slug, description, imageUrl, sortOrder, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
...
    GalleryItem:
      type: object
      required: [id, title, description, imageUrl, sortOrder, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
    GalleryInput:
      type: object
      required: [title, imageUrl]
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: string }
        sortOrder: { type: integer }
    Inquiry:
      type: object
      required: [id, name, phone, email, message, items, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        message: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/InquiryItem" }
        status: { type: string, enum: [new, contacted, closed] }
        createdAt: { type: string, format: date-time }
    InquiryItem:
      type: object
      required: [productId, productName, quantity]
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
    InquiryInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        message: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/InquiryItem" }
    InquiryUpdate:
      type: object
      properties:
        status: { type: string, enum: [new, contacted, closed] }
...