# shared/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Qattan Curtains showroom API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItem {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  imageUrl: string;
  sortOrder: number;
  createdAt: Date;
}