# shared/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Qattan Curtains showroom API
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryItem } from "./inquiryItem";
import type { InquiryStatus } from "./inquiryStatus";

export interface Inquiry {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  message: string | null;
  items: InquiryItem[];
  status: InquiryStatus;
  createdAt: Date;
}